DTK-952 When dtk shell started from specific module/service folder and that module/service folder delted, re-import from remote will fail

+++ lib/commands/thor/module.rb
@@ -167,6 +167,14 @@ module DTK::Client
         modules_path        = OsUtil.module_clone_location()
         module_location     = "#{modules_path}/#{component_module_id}" unless component_module_id.nil?
         module_location     = module_location + "-#{version}" if options.version?
+        pwd                 = Dir.getwd()
+
+        if (pwd == module_location)
+          DTK::Client::OsUtil.print("Local directory '#{module_location}' is not deleted because it is your current working directory.", :yellow) 
+          
+          puts "You have successfully deleted component module '#{component_module_id}'."
+          return response
+        end
 
         FileUtils.rm_rf("#{module_location}") if (File.directory?(module_location) && ("#{modules_path}/" != module_location))
